requirements.txt
ultralytics==8.0.206
torch==2.1.0         # or any version that works with your setup and model
torchvision==0.16.0
opencv-python
Pillow
mediapipe
numpy
